Set up build configuration for x64 and updated the project to cast types to suppress type conversion warnings. The options menu now supports detection of mouse over menu items as well as selecting items.

+++ Menu2/NetCode.cpp
@@ -95,7 +95,7 @@ void NetworkPost(LPSTR url, LPSTR host, LPSTR data)
 		return;
 	}
 
-	send(Socket, packet.str().c_str(), packet.str().length(), 0);
+	send(Socket, packet.str().c_str(), (int)packet.str().length(), 0);
 
 	while (true)
 	{
@@ -156,7 +156,7 @@ void NetworkGet(LPSTR url, LPSTR host, LPSTR data)
 		return;
 	}
 
-	send(Socket, packet.str().c_str(), packet.str().length(), 0);
+	send(Socket, packet.str().c_str(), (int)packet.str().length(), 0);
 
 	while (true)
 	{
